requirements.txt
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.6
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-Login==0.5.0
flask-mongoengine==1.0.0
Flask-PyMongo==2.3.0
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.10
is-disposable-email==1.0.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
monai @ git+https://github.com/Project-MONAI/MONAI@2eada01efd9ddbcf8d9befe78b6ed67ce0cb454f
mongoengine==0.23.0
numpy==1.20.1
Pillow==8.1.2
pycparser==2.20
pymongo==3.11.3
pyOpenSSL==20.0.1
python-dotenv==0.16.0
requests==2.25.1
six==1.15.0
torch==1.8.0
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
WTForms==2.3.3
wtforms-validators==1.0.0
